Fix/errosApi

+++ app/Http/Controllers/Api/ProductController.php
@@ -20,7 +20,13 @@ class ProductController extends Controller
      */
     public function index()
     {
-        return $this->product->paginate(10);
+        $product = $this->product->paginate(10);
+
+        if ($product) {
+            return $product;
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 
     /**
@@ -28,7 +34,13 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        return $this->product->create($request->all());
+        $product =  $this->product->create($request->all());
+
+        if ($product) {
+            return $product;
+        } else {
+            return response()->json(['error' => 'Erro criação'], 500);
+        }
     }
 
     /**
@@ -36,7 +48,13 @@ class ProductController extends Controller
      */
     public function show(string $id)
     {
-        return $this->product->find($id);
+        $product =  $this->product->find($id);
+
+        if ($product) {
+            return $product;
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 
     /**
@@ -45,9 +63,12 @@ class ProductController extends Controller
     public function update(Request $request, string $id)
     {
         $product = $this->product->find($id);
-        $product->update($request->all());
 
-        return $product;
+        if ($product) {
+            return $product->update($request->all());
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 
     /**
@@ -56,8 +77,11 @@ class ProductController extends Controller
     public function destroy(string $id)
     {
         $product = $this->product->find($id);
-        $product->delete();
 
-        return $product;
+        if ($product) {
+            return $product->delete();
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 }
