+++ app/Http/Controllers/Api/UserController.php
@@ -21,7 +21,13 @@ class UserController extends Controller
      */
     public function index()
     {
-        return $this->user->paginate(10)->load('clients');
+        $user = $this->user->paginate(10);
+
+        if ($user) {
+            return $user;
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 
     /**
@@ -37,7 +43,13 @@ class UserController extends Controller
         //     'remember_token' => Str::random(10),
         // ]);
 
-        return $this->user->create($request->all())->load('clients');
+        $user =  $this->user->create($request->all());
+
+        if ($user) {
+            return $user->load('clients');
+        } else {
+            return response()->json(['error' => 'Erro criação'], 500);
+        }
     }
 
     /**
@@ -45,7 +57,13 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
-        return $this->user->find($id)->load('clients');
+        $user = $this->user->find($id);
+
+        if ($user) {
+            return $user->load('clients');
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 
     /**
@@ -53,10 +71,14 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $user = $this->user->find($id)->load('clients');
-        $user->update($request->all());
+        $user = $this->user->find($id);
 
-        return $user;
+        if ($user) {
+            $user->load('clients');
+            return $user->update($request->all());
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 
     /**
@@ -64,9 +86,13 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
-        $user = $this->user->find($id)->load('clients');
-        $user->delete();
+        $user = $this->user->find($id);
 
-        return $user;
+        if ($user) {
+            $user->load('clients');
+            return $user->delete();
+        } else {
+            return response()->json(['error' => 'Not found'], 404);
+        }
     }
 }
